+++ src/main/kotlin/stringcalculator/domain/customparser/ParserSeparator.kt
@@ -1,4 +1,4 @@
-package stringcalculator.domain
+package stringcalculator.domain.customparser
 
 class ParserSeparator(val string: String) {
