+++ src/main/kotlin/stringcalculator/service/CustomExpressionCalculator.kt
@@ -1,14 +1,14 @@
 package stringcalculator.service
 
 import stringcalculator.domain.Calculator
-import stringcalculator.domain.StringExpressionParser
+import stringcalculator.domain.customparser.CustomExpressionParser
 import stringcalculator.dto.InputDto
 import stringcalculator.dto.OutPutDto
 
 class CustomExpressionCalculator(inputDto: InputDto) {
-    private val stringExpressionParser: StringExpressionParser = StringExpressionParser(inputDto.inputString)
+    private val customExpressionParser: CustomExpressionParser = CustomExpressionParser(inputDto.inputString)
 
     fun getTotal(): OutPutDto {
-        return OutPutDto(Calculator.getTotalSum(stringExpressionParser.getParsedNumber()))
+        return OutPutDto(Calculator.getTotalSum(customExpressionParser.getParsedNumber()))
     }
 }
