+++ src/test/kotlin/stringcalculator/domain/customparser/ParserSeparatorTest.kt
@@ -1,4 +1,4 @@
-package stringcalculator.domain
+package stringcalculator.domain.customparser
 
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.DescribeSpec
